Графический процесс: добавлен gef.EditPart (aurusov/raox#334)
