+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/process/ProcessEditPart.java
@@ -1,7 +1,6 @@
 package ru.bmstu.rk9.rao.ui.process;
 
 import java.beans.PropertyChangeEvent;
-import java.beans.PropertyChangeListener;
 import java.util.List;
 
 import org.eclipse.draw2d.geometry.Point;
@@ -10,7 +9,6 @@ import org.eclipse.gef.EditPolicy;
 import org.eclipse.gef.NodeEditPart;
 import org.eclipse.gef.Request;
 import org.eclipse.gef.RequestConstants;
-import org.eclipse.gef.editparts.AbstractGraphicalEditPart;
 import org.eclipse.gef.requests.DropRequest;
 import org.eclipse.swt.graphics.Color;
 import org.eclipse.swt.graphics.RGB;
@@ -19,38 +17,26 @@ import org.eclipse.ui.IWorkbenchPage;
 import org.eclipse.ui.PartInitException;
 import org.eclipse.ui.PlatformUI;
 
+import ru.bmstu.rk9.rao.ui.gef.EditPart;
 import ru.bmstu.rk9.rao.ui.process.connection.Connection;
 import ru.bmstu.rk9.rao.ui.process.node.BlockNode;
 import ru.bmstu.rk9.rao.ui.process.node.Node;
 
-public abstract class ProcessEditPart extends AbstractGraphicalEditPart
-		implements PropertyChangeListener, NodeEditPart {
+public abstract class ProcessEditPart extends EditPart implements NodeEditPart {
 
 	private int ID;
 
-	public void setID(int ID) {
-		this.ID = ID;
-	}
-	
 	public int getID() {
 		return ID;
+	}
+
+	public void setID(int ID) {
+		this.ID = ID;
 	}
 
 	@Override
 	public final List<ru.bmstu.rk9.rao.ui.gef.Node> getModelChildren() {
 		return ((ru.bmstu.rk9.rao.ui.gef.Node) getModel()).getChildren();
-	}
-
-	@Override
-	public void activate() {
-		super.activate();
-		((Node) getModel()).addPropertyChangeListener(this);
-	}
-
-	@Override
-	public void deactivate() {
-		super.deactivate();
-		((Node) getModel()).removePropertyChangeListener(this);
 	}
 
 	@Override
@@ -67,37 +53,38 @@ public abstract class ProcessEditPart extends AbstractGraphicalEditPart
 
 	@Override
 	public void propertyChange(PropertyChangeEvent evt) {
-		if (evt.getPropertyName().equals(Node.PROPERTY_COLOR)) {
-			getFigure().setBackgroundColor(new Color(null, (RGB) evt.getNewValue()));
-		}
-		if (evt.getPropertyName().equals(Node.PROPERTY_NAME)) {
+		super.propertyChange(evt);
+
+		if (evt.getPropertyName().equals(Node.PROPERTY_NAME))
 			((ProcessFigure) getFigure()).setFigureNameVisible((boolean) evt.getNewValue());
-		}
-		if (evt.getPropertyName().equals(Node.PROPERTY_CONSTRAINT))
-			refreshVisuals();
+
 		if (evt.getPropertyName().equals(BlockNode.SOURCE_CONNECTION_UPDATED))
 			refreshSourceConnections();
+
 		if (evt.getPropertyName().equals(BlockNode.TARGET_CONNECTION_UPDATED))
 			refreshTargetConnections();
 	}
 
 	@Override
 	protected void createEditPolicies() {
-		installEditPolicy(EditPolicy.COMPONENT_ROLE, new ProcessDeletePolicy());
+		super.createEditPolicies();
 		installEditPolicy(EditPolicy.GRAPHICAL_NODE_ROLE, new ConnectionPolicy());
 	}
 
 	@Override
 	protected void refreshVisuals() {
+		super.refreshVisuals();
+
 		if (!(getFigure() instanceof ProcessFigure))
 			return;
+
 		ProcessFigure figure = (ProcessFigure) getFigure();
 		Node node = (Node) getModel();
 
-		figure.setConstraint(node.getConstraint());
 		RGB oldColor = node.getColor();
 		Color newColor = new Color(null, oldColor);
 		figure.setBackgroundColor(newColor);
+
 		figure.setFigureNameVisible(node.nameIsVisible());
 	}
